+++ player/ControllableCamera.gd
@@ -2,49 +2,35 @@ extends Spatial
 
 class_name ControllableCamera
 
-export(float) var min_pitch: float = -90
-export(float) var max_pitch: float = 75
 export(float) var rotate_speed: float = 10
-export(float) var sensitivity: float = 0.1
 export(float) var zoom_speed: float = 10
-export(float) var zoom_step: float = .5
 export(float) var min_distance: float = 3
 export(float) var max_distance: float = 10
 
 onready var _gimbal_h: Spatial = $GimbalH
 onready var _gimbal_v: Spatial = $GimbalH/GimbalV
 onready var _camera: ClippedCamera = $GimbalH/GimbalV/ClippedCamera
+onready var _controls: Controls = get_tree().get_nodes_in_group("controls")[0]
 
 var _rot_h: float = 0
 var _rot_v: float = 0
-var _is_capturing: bool = true
 var _distance: float = 0
 
 func _ready():
     yield(get_parent(), "ready")
     _distance = _camera.transform.origin.z
-    Input.set_mouse_mode(Input.MOUSE_MODE_CAPTURED)
+    call_deferred("_initialize_zoom_scale")
 
-func _input(event):
-    if Input.is_action_just_pressed("ui_cancel"):
-        _is_capturing = !_is_capturing
+func _initialize_zoom_scale():
+    var initial_zoom_scale := (_distance - min_distance) / (max_distance - min_distance)
+    _controls.set_zoom_scale(initial_zoom_scale)
 
-        if _is_capturing:
-            Input.set_mouse_mode(Input.MOUSE_MODE_CAPTURED)
-        else:
-            Input.set_mouse_mode(Input.MOUSE_MODE_VISIBLE)
+func _process(delta):
+    var cam_rot: Vector2 = _controls.get_camera_rotation()
+    _rot_h = cam_rot.x
+    _rot_v = cam_rot.y
 
-    if _is_capturing && event is InputEventMouseMotion:
-        _rot_h -= event.relative.x * sensitivity
-        _rot_v -= event.relative.y * sensitivity
-
-    _rot_v = clamp(_rot_v, min_pitch, max_pitch)
-
-    if _is_capturing:
-        if Input.is_action_just_pressed("zoom_in"):
-            _distance = clamp(_distance - zoom_step, min_distance, max_distance)
-        if Input.is_action_just_pressed("zoom_out"):
-            _distance = clamp(_distance + zoom_step, min_distance, max_distance)
+    _distance = _controls.get_zoom_scale() * (max_distance - min_distance) + min_distance
 
 func _physics_process(delta):
     _gimbal_h.rotation_degrees.y = lerp(_gimbal_h.rotation_degrees.y, _rot_h, rotate_speed * delta)